+++ T02/S04T02N02GrauHortaNuria/src/main/java/cat/itacademy/barcelonactiva/grauhorta/nuria/s04/t02/n02/model/repository/FruitRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface FruitRepository extends JpaRepository<Fruit, Long> {
+public interface FruitRepository extends JpaRepository<Fruit, Integer> {
 }
